refactor : LottoResults 클래스 생성 및 전체적인 구조 수정
